[dashboard] Simplify SettingsPage component by folding the SubMenu back into it
