+++ components/dashboard/src/settings/EnvironmentVariables.tsx
@@ -1,9 +1,8 @@
-import { SettingsPage } from "../components/SettingsPage";
-import settingsMenu from './settings-menu';
+import { SettingsPage } from "./SettingsPage";
 
 export default function EnvVars() {
     return <div>
-        <SettingsPage title='Environment Variables' subtitle='Configure environment variables for your workspaces' menuEntries={settingsMenu}>
+        <SettingsPage title='Environment Variables' subtitle='Configure environment variables for your workspaces'>
             <h3>Environment Variables</h3>
         </SettingsPage>
     </div>;
