+++ components/dashboard/src/settings/GitIntegrations.tsx
@@ -1,9 +1,8 @@
-import { SettingsPage } from "../components/SettingsPage";
-import settingsMenu from './settings-menu';
+import { SettingsPage } from "./SettingsPage";
 
 export default function GitIntegrations() {
     return <div>
-        <SettingsPage title='Git Integrations' subtitle='Manage integration with your Git hosters' menuEntries={settingsMenu}>
+        <SettingsPage title='Git Integrations' subtitle='Manage integration with your Git hosters'>
             <h3>Git Hoster Access Control</h3>
         </SettingsPage>
     </div>;
