+++ components/dashboard/src/settings/Plans.tsx
@@ -1,9 +1,8 @@
-import { SettingsPage } from "../components/SettingsPage";
-import settingsMenu from "./settings-menu";
+import { SettingsPage } from "./SettingsPage";
 
 export default function Plans() {
     return <div>
-        <SettingsPage title='Plans' subtitle='Plans and Usage' menuEntries={settingsMenu} >
+        <SettingsPage title='Plans' subtitle='Plans and Usage'>
             <h3>Plans</h3>
         </SettingsPage>
     </div>;
